Add a DAHDI hook script for Asterisk
This script is distributed with dahdi-tools but should only be installed
if we have Asterisk installed.

When a new DAHDI span is created, dahdi runs several hook scripts (from
udev). This one is among them. It does nothing if there's no DAHDI
device on the system and hence safe to add to the main package.

Closes #848584.